requirements.txt
# Scraping
selenium>=4.15
webdriver-manager>=4.0
beautifulsoup4>=4.12
requests>=2.31
tqdm>=4.66

# ML
catboost>=1.2
xgboost>=2.0
scikit-learn>=1.3
optuna>=3.4
shap>=0.43
numpy>=1.24
pandas>=2.1

# Images
Pillow>=10.0

# Dashboard
streamlit>=1.29
plotly>=5.18

# Utils
thefuzz>=0.20
python-Levenshtein>=0.23
python-dotenv>=1.0
